requirements.txt
instaloader>=4.10
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
